Use the BreezSDK dependency to manage notification jobs
